+++ plugins/sitemap.ts
@@ -78,8 +78,9 @@ export function sitemap(userOptions?: Options) {
 
       // Add or update `robots.txt` with the sitemap url
       const robots = await site.getOrCreatePage("/robots.txt");
-      const content = robots.text || `User-agent: *\nAllow: /\n`;
-      robots.text = `${content}\nSitemap: ${site.url(options.filename, true)}`;
+      robots.text = `${robots.text}\nSitemap: ${
+        site.url(options.filename, true)
+      }`;
     });
 
     function generateSitemap(pages: Data[]): string {